+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # CHANGELOG
+### Version 6.0
+- Redirect to the previous page if session is expired
+- Fixes errors with PHP8
+- Better stock management when making a formula
+- Add ingredient info in formula making  view
+- Add schedule confirmation and explanation message
+- Add ingredient info in formula making screen
+- Add option to create a new formula category from the formulas list page
+- Add low stock in formula making screen
+- Improve ingredient replacement management
+- Add stock value in ingredient replacement suggestion
+
 ### Version 5.9
-- Introducing formulas import/exort in JSON format
+- Introducing formulas import/export in JSON format
 - Notify but don't prevent formula listing if no ingredients in database
 - Added option to exclude/include synonyms and compositions when importing from PV Online
 - Bash run script update to utilise docker volumes